README.md: Add more q&a and edit `Is it safe?` in `Possible questions`, shorten list of dependencies.
Add `Which DE or WM should I use for best compatibility?`, `Does it increase input lag and/or record my screen to detect window focus/unfocus events?` and `Is not running commands on focus and unfocus makes system vulnerable?` to `Possible questions` section.
Shorten list of dependencies in `Dependencies section`.